// model/command.ts
/**
 * QuantCDN Unified API
 * Unified API for QuantCDN Admin and QuantCloud Platform services
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class Command {
    'runId'?: string;
    'runType'?: string;
    'command'?: string;
    'status'?: string;
    'startTime'?: string;
    'endTime'?: string;
    'exitCode'?: number;
    'output'?: Array<string>;
    'scheduleName'?: string;
    'targetContainerName'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "runId",
            "baseName": "runId",
            "type": "string"
        },
        {
            "name": "runType",
            "baseName": "runType",
            "type": "string"
        },
        {
            "name": "command",
            "baseName": "command",
            "type": "string"
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "string"
        },
        {
            "name": "startTime",
            "baseName": "startTime",
            "type": "string"
        },
        {
            "name": "endTime",
            "baseName": "endTime",
            "type": "string"
        },
        {
            "name": "exitCode",
            "baseName": "exitCode",
            "type": "number"
        },
        {
            "name": "output",
            "baseName": "output",
            "type": "Array<string>"
        },
        {
            "name": "scheduleName",
            "baseName": "scheduleName",
            "type": "string"
        },
        {
            "name": "targetContainerName",
            "baseName": "targetContainerName",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return Command.attributeTypeMap;
    }
}