// model/containerMountPointsInner.ts
/**
 * QuantCDN Unified API
 * Unified API for QuantCDN Admin and QuantCloud Platform services
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class ContainerMountPointsInner {
    'sourceVolume': string;
    'containerPath': string;
    'readOnly'?: boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "sourceVolume",
            "baseName": "sourceVolume",
            "type": "string"
        },
        {
            "name": "containerPath",
            "baseName": "containerPath",
            "type": "string"
        },
        {
            "name": "readOnly",
            "baseName": "readOnly",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return ContainerMountPointsInner.attributeTypeMap;
    }
}