// model/createCommandRequest.ts
/**
 * QuantCloud API
 * QuantCloud API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CreateCommandRequest {
    'command'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "command",
            "baseName": "command",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CreateCommandRequest.attributeTypeMap;
    }
}