// model/createCronJobRequest.ts
/**
 * QuantCloud API
 * QuantCloud API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CreateCronJobRequest {
    'name': string;
    'description'?: string | null;
    'scheduleExpression': string;
    'command': Array<string>;
    'targetContainerName'?: string | null;
    'isEnabled'?: boolean | null = true;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "scheduleExpression",
            "baseName": "scheduleExpression",
            "type": "string"
        },
        {
            "name": "command",
            "baseName": "command",
            "type": "Array<string>"
        },
        {
            "name": "targetContainerName",
            "baseName": "targetContainerName",
            "type": "string"
        },
        {
            "name": "isEnabled",
            "baseName": "isEnabled",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return CreateCronJobRequest.attributeTypeMap;
    }
}